+++ src/main/java/io/socket/client/Ack.java
@@ -5,5 +5,5 @@ package io.socket.client;
  */
 
 public interface Ack {
-    public void call(Object... args);
+    void call(Object... args);
 }