+++ src/main/java/io/socket/client/Url.java
@@ -1,23 +1,29 @@
 package io.socket.client;
 
-import java.net.MalformedURLException;
 import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.URL;
+import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class Url {
 
-    private static Pattern PATTERN_HTTP = Pattern.compile("^http|ws$");
-    private static Pattern PATTERN_HTTPS = Pattern.compile("^(http|ws)s$");
+    /**
+     * Expected format: "[id:password@]host[:port]"
+     */
+    private static Pattern PATTERN_AUTHORITY = Pattern.compile("^(.*@)?([^:]+)(:\\d+)?$");
 
     private Url() {}
 
-    public static URL parse(String uri) throws URISyntaxException {
-        return parse(new URI(uri));
+    static class ParsedURI {
+        public final URI uri;
+        public final String id;
+
+        public ParsedURI(URI uri, String id) {
+            this.uri = uri;
+            this.id = id;
+        }
     }
 
-    public static URL parse(URI uri) {
+    public static ParsedURI parse(URI uri) {
         String protocol = uri.getScheme();
         if (protocol == null || !protocol.matches("^https?|wss?$")) {
             protocol = "https";
@@ -25,9 +31,9 @@ public class Url {
 
         int port = uri.getPort();
         if (port == -1) {
-            if (PATTERN_HTTP.matcher(protocol).matches()) {
+            if ("http".equals(protocol) || "ws".equals(protocol)) {
                 port = 80;
-            } else if (PATTERN_HTTPS.matcher(protocol).matches()) {
+            } else if ("https".equals(protocol) || "wss".equals(protocol)) {
                 port = 443;
             }
         }
@@ -40,34 +46,39 @@ public class Url {
         String userInfo = uri.getRawUserInfo();
         String query = uri.getRawQuery();
         String fragment = uri.getRawFragment();
-        try {
-            return new URL(protocol + "://"
-                    + (userInfo != null ? userInfo + "@" : "")
-                    + uri.getHost()
-                    + (port != -1 ? ":" + port : "")
-                    + path
-                    + (query != null ? "?" + query : "")
-                    + (fragment != null ? "#" + fragment : ""));
-        } catch (MalformedURLException e) {
-            throw new RuntimeException(e);
+        String _host = uri.getHost();
+        if (_host == null) {
+            // might happen on some of Samsung Devices such as S4.
+            _host = extractHostFromAuthorityPart(uri.getRawAuthority());
         }
+        URI completeUri = URI.create(protocol + "://"
+                + (userInfo != null ? userInfo + "@" : "")
+                + _host
+                + (port != -1 ? ":" + port : "")
+                + path
+                + (query != null ? "?" + query : "")
+                + (fragment != null ? "#" + fragment : ""));
+        String id = protocol + "://" + _host + ":" + port;
+
+        return new ParsedURI(completeUri, id);
     }
 
-    public static String extractId(String url) throws MalformedURLException {
-        return extractId(new URL(url));
-    }
 
-    public static String extractId(URL url) {
-        String protocol = url.getProtocol();
-        int port = url.getPort();
-        if (port == -1) {
-            if (PATTERN_HTTP.matcher(protocol).matches()) {
-                port = 80;
-            } else if (PATTERN_HTTPS.matcher(protocol).matches()) {
-                port = 443;
-            }
+    private static String extractHostFromAuthorityPart(String authority)
+    {
+        if (authority == null) {
+            throw new RuntimeException("unable to parse the host from the authority");
         }
-        return protocol + "://" + url.getHost() + ":" + port;
+
+        Matcher matcher = PATTERN_AUTHORITY.matcher(authority);
+
+        // If the authority part does not match the expected format.
+        if (!matcher.matches()) {
+            throw new RuntimeException("unable to parse the host from the authority");
+        }
+
+        // Return the host part.
+        return matcher.group(2);
     }
 
 }