+++ src/main/java/io/socket/parser/Parser.java
@@ -5,44 +5,44 @@ public interface Parser {
     /**
      * Packet type `connect`.
      */
-    public static final int CONNECT = 0;
+    int CONNECT = 0;
 
     /**
      * Packet type `disconnect`.
      */
-    public static final int DISCONNECT = 1;
+    int DISCONNECT = 1;
 
     /**
      * Packet type `event`.
      */
-    public static final int EVENT = 2;
+    int EVENT = 2;
 
     /**
      * Packet type `ack`.
      */
-    public static final int ACK = 3;
+    int ACK = 3;
 
     /**
      * Packet type `error`.
      */
-    public static final int ERROR = 4;
+    int CONNECT_ERROR = 4;
 
     /**
      * Packet type `binary event`.
      */
-    public static final int BINARY_EVENT = 5;
+    int BINARY_EVENT = 5;
 
     /**
      * Packet type `binary ack`.
      */
-    public static final int BINARY_ACK = 6;
+    int BINARY_ACK = 6;
 
-    public static int protocol = 4;
+    int protocol = 5;
 
     /**
      * Packet types.
      */
-    public static String[] types = new String[] {
+    String[] types = new String[] {
         "CONNECT",
         "DISCONNECT",
         "EVENT",
@@ -52,29 +52,29 @@ public interface Parser {
         "BINARY_ACK"
     };
 
-    public static interface Encoder {
+    interface Encoder {
 
-        public void encode(Packet obj, Callback callback);
+        void encode(Packet obj, Callback callback);
 
-        public interface Callback {
+        interface Callback {
 
-            public void call(Object[] data);
+            void call(Object[] data);
         }
     }
 
-    public static interface Decoder {
+    interface Decoder {
 
-        public void add(String obj);
+        void add(String obj);
 
-        public void add(byte[] obj);
+        void add(byte[] obj);
 
-        public void destroy();
+        void destroy();
 
-        public void onDecoded(Callback callback);
+        void onDecoded(Callback callback);
 
-        public interface Callback {
+        interface Callback {
 
-            public void call(Packet packet);
+            void call(Packet packet);
         }
     }
 }