+++ src/test/java/io/socket/parser/Helpers.java
@@ -7,11 +7,11 @@ import org.skyscreamer.jsonassert.JSONAssert;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
+import static org.junit.Assert.fail;
 
 public class Helpers {
 
     private static Parser.Encoder encoder = new IOParser.Encoder();
-    private static Packet<String> errorPacket = new Packet<String>(Parser.ERROR, "parser error");
 
     public static void test(final Packet obj) {
         encoder.encode(obj, new Parser.Encoder.Callback() {
@@ -31,13 +31,10 @@ public class Helpers {
 
     public static void testDecodeError(final String errorMessage) {
         Parser.Decoder decoder = new IOParser.Decoder();
-        decoder.onDecoded(new IOParser.Decoder.Callback() {
-            @Override
-            public void call(Packet packet) {
-                assertPacket(errorPacket, packet);
-            }
-        });
-        decoder.add(errorMessage);
+        try {
+            decoder.add(errorMessage);
+            fail();
+        } catch (DecodingException e) {}
     }
 
     @SuppressWarnings("unchecked")